+++ src/mesh/core/result.clj
@@ -7,6 +7,11 @@
   (when (empty? @global_graph_set)
     (reset! global_graph_set (persist/read-graph-names data_dir)))
   (cond
+    (not-empty (re-matches quit_re raw_query_string))
+    (do
+      (print " ")
+      (when (contains? @global_graph_set @global_graph_name) (persist/write-graph @global_graph data_dir))
+      (do (println) (System/exit 0)))
     (not-empty (re-matches list_graphs_re raw_query_string))
     (do
       (print " ")
@@ -28,31 +33,31 @@
     (if (not (contains? @global_graph_set (nth (str/split (re-find create_graph_re raw_query_string) #"\s{1,}") 1)))
       (do
         (print " ")
-        (execute-create-graph-query (subs raw_query_string 6)))
+        (create-graphs (subs raw_query_string 6)))
       (print " Graph already created."))
     (not-empty (re-matches delete_graph_re raw_query_string))
     (if (contains? @global_graph_set (nth (str/split (re-find delete_graph_re raw_query_string) #"\s{1,}") 1))
       (do
         (print " ")
-        (execute-delete-graph-query (subs raw_query_string 6)))
+        (delete-graphs (subs raw_query_string 6)))
       (print " Graph not found."))
     (not-empty (re-matches create_node_re raw_query_string))
     (if (contains? @global_graph_set @global_graph_name)
       (do
         (print " ")
-        (spin/spin! #(execute-create-node-query @global_graph_name (subs raw_query_string 6)) {:frames (spin/styles :braille)}))
+        (spin/spin! #(create-nodes @global_graph_name (subs raw_query_string 6)) {:frames (spin/styles :braille)}))
       (print " Graph not found."))
     (not-empty (re-matches show_node_re raw_query_string))
     (if (contains? @global_graph_set @global_graph_name)
       (do
         (print " ")
-        (pprint/print-table (spin/spin! #(execute-show-node-query (subs raw_query_string 4)) {:frames (spin/styles :braille)})))
+        (pprint/print-table (spin/spin! #(show-nodes (subs raw_query_string 4)) {:frames (spin/styles :braille)})))
       (print " Graph not found."))
     (not-empty (re-matches create_edge_re raw_query_string))
     (if (contains? @global_graph_set @global_graph_name)
       (do
         (print " ")
-        (spin/spin! #(execute-create-edge-query @global_graph_name (subs raw_query_string 6)) {:frames (spin/styles :braille)}))
+        (spin/spin! #(create-edges @global_graph_name (subs raw_query_string 6)) {:frames (spin/styles :braille)}))
       (print " Graph not found."))
     :else
     (print ""))